Organized constants in Utility.h

+++ ThrowIT/ViewComponents/PartyCell.m
@@ -28,7 +28,7 @@
         self.party.numberAttending -= 1;
         [self.party saveInBackground];
 
-        self.partyGoingCount.text = [NSString stringWithFormat:@"%d", self.party.numberAttending];
+        self.partyGoingCount.text = [NSString stringWithFormat:@"%ld", (long)self.party.numberAttending];
         
     }
     else if([self.goingButton.titleLabel.text isEqualToString:@"Maybe"]){
@@ -38,7 +38,7 @@
         [self.goingButton setTitle:@"Going" forState:UIControlStateNormal];
         self.party.numberAttending += 1;
         [self.party saveInBackground];
-        self.partyGoingCount.text = [NSString stringWithFormat:@"%d", self.party.numberAttending];
+        self.partyGoingCount.text = [NSString stringWithFormat:@"%ld", (long)self.party.numberAttending];
     }
 }
